reaction cooldown, mute save, optimizing

+++ event/on_message.py
@@ -1,53 +1,68 @@
 import random
-
 import discord
-
+import time
 from emoji.emoji import delete_emoji
 from iBot import client, emoji_list, muted_list
+from mute.mute import delMute
+from reaction.reaction import getReactionTime, setReactionTime
 from strings.quotes import quotes, answers
 
 
 @client.event
 async def on_message(message):
-    if message.author == client.user:
+    # we dont handle anything from ibot itself
+    if message.author.id == client.user.id:
         return
-    for x in muted_list:
-        if x.user == message.author.id:
-            if x.flag == 1:
-                try:
-                    await message.delete()
-                except:
-                    continue
-            else:
-                if x.server == message.channel.guild.id:
-                    try:
-                        await message.delete()
-                    except:
-                        continue
 
     if not isinstance(message.channel, discord.abc.PrivateChannel):
-        for x in emoji_list:
-            if message.channel.guild.id == x.server and message.author.id == x.user:
-                emoji = client.get_emoji(x.emoji)
+        await handleMute(message)
+        await handleReaction(message)
+        if ('cj' in message.content.lower() or "ibot" in message.content.lower() or "ilearner" in message.content.lower()) and not message.author.bot:
+            await handleMessage(message)
+
+        if message.mentions:
+            await handleMention(message)
+
+    if isinstance(message.channel, discord.abc.PrivateChannel):
+        await handleDm(message)
+
+    await client.process_commands(message)
+
+
+async def handleMute(message):
+    for x in muted_list:
+        if x.user == message.author.id and x.server == message.channel.guild.id:
+            try:
+                await message.delete()
+            except:
+                await delMute(message.author.id, message.channel.guild.id)
+
+
+async def handleReaction(message):
+
+    for x in emoji_list:
+        if message.channel.guild.id == x.server and message.author.id == x.user:
+            cur_time = int(time.time())
+            last_reaction_time = await getReactionTime(message.author.id)
+            if (cur_time - last_reaction_time) > 2:
                 try:
+                    emoji = client.get_emoji(x.emoji)
                     await message.add_reaction(emoji)
+                    setReactionTime(message.author.id, cur_time)
                 except discord.errors.HTTPException:
                     await delete_emoji(emoji.id, x.user, message.channel.guild.id)
 
-    if isinstance(message.channel, discord.abc.PrivateChannel):
-        item = random.choice(quotes)
-        await message.channel.send(item)
 
-    if (
-            "cj" in message.content.lower() or "ibot" in message.content.lower() or "ilearner" in message.content.lower()) and message.author.id != 555845842662064128 and not message.author.bot:
-        await handleMessage(message)
+async def handleMention(message):
+    for member in message.mentions:
+        if member.id == 379624030317182976:
+            item = random.choice(quotes)
+            await message.channel.send(item)
 
-    if message.mentions:
-        for member in message.mentions:
-            if member.id == 379624030317182976:
-                item = random.choice(quotes)
-                await message.channel.send(item)
-    await client.process_commands(message)
+
+async def handleDm(message):
+    item = random.choice(quotes)
+    await message.channel.send(item)
 
 
 async def handleMessage(message):
